refactor(web): authenticateのロジック置き場をlib/からmodel/へ
汎用共有を連想させるlibは名前負けのため、feature固有ロジックを表すmodelに改める。
