+++ application/packages/web/src/client/features/authenticate/index.ts
@@ -3,8 +3,8 @@ export {
   AUTH_ERROR_MESSAGES,
   resolveLoginErrorMessage,
   resolveSignupErrorMessage,
-} from './lib/error-message'
-export { resolveTurnstileSiteKey } from './lib/turnstile'
+} from './model/error-message'
+export { resolveTurnstileSiteKey } from './model/turnstile'
 export { default as useLogin } from './hooks/use-login'
 export { default as useSession, SESSION_SWR_KEY } from './hooks/use-session'
 export { default as useSignup } from './hooks/use-signup'
@@ -12,4 +12,4 @@ export {
   type AuthenticateErrors,
   type AuthenticateFormData,
   validateAuthenticateForm,
-} from './lib/validation'
+} from './model/validation'
